More meaningless twiddling.  This isn't working, I need a GPU debugger I think

+++ src/shape.hpp
@@ -56,7 +56,7 @@ enum AFK_SkeletonFlag
 class AFK_SkeletonFlagGrid
 {
 protected:
-    unsigned int **grid;
+    unsigned long long **grid;
     int gridDim;
 
 public:
